crud and input warehouse by hhp

+++ giftstore_website/database/migrations/2023_04_06_145937_create_bill_order_details_table.php
@@ -18,7 +18,7 @@ class CreateBillOrderDetailsTable extends Migration
             $table->string('id_bill_order'); //foreign key
             $table->string('id_product'); //foreign key
             $table->integer('quantity')->default(0);
-            $table->float('purchase_price')->default(0);
+            $table->bigInteger('purchase_price')->default(0);
             $table->bigInteger('total_price')->default(0);
             $table->string('status')->default('enabled');
             $table->timestamps();
